requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

colossalai==0.2.8
einops==0.6.0
hydra-core==1.3.2
lightning==2.0.1
lightning-colossalai==0.1.0
rich==13.3.3
torch==1.13.1
fastapi==0.88.0
tensorboard==2.12.1
torchaudio==0.13.1
transformers==4.27.4
